Close #106 - Re-organize data model
