+++ modules/openai4s-core/shared/src/main/scala-2/openai4s/types/chat/Message.scala
@@ -1,4 +1,4 @@
-package openai4s.types
+package openai4s.types.chat
 
 import cats.{Eq, Show}
 import extras.render.Render
