+++ modules/openai4s-core/shared/src/main/scala-2/openai4s/types/completions/Text.scala
@@ -1,17 +1,15 @@
 package openai4s.types.completions
 
-import cats.syntax.all.*
 import cats.{Eq, Show}
-import eu.timepit.refined.api.{Refined, RefinedTypeOps}
 import eu.timepit.refined.cats.*
-import eu.timepit.refined.numeric.Interval
-import eu.timepit.refined.types.numeric.{NonNegFloat, PosInt}
+import eu.timepit.refined.types.numeric.NonNegFloat
 import eu.timepit.refined.types.string.NonEmptyString
 import extras.render.Render
 import io.circe.generic.extras.Configuration
 import io.circe.refined.*
 import io.circe.{Decoder, Encoder}
 import io.estatico.newtype.macros.newtype
+import openai4s.types.common.*
 
 /** https://platform.openai.com/docs/api-reference/completions#:~:text=prompt%20and%20parameters.-,Request%20body,-model
   * @author Kevin Lee
@@ -20,8 +18,8 @@ import io.estatico.newtype.macros.newtype
 final case class Text(
   model: Model,
   prompt: Option[Text.Prompt],
-  maxTokens: Option[Text.MaxTokens],
-  temperature: Option[Text.Temperature],
+  maxTokens: Option[MaxTokens],
+  temperature: Option[Temperature],
   topP: Option[Text.TopP],
   n: Option[Text.N],
   stream: Option[Text.Stream],
@@ -50,27 +48,6 @@ object Text {
 
     implicit val promptEncoder: Encoder[Prompt] = deriving
     implicit val promptDecoder: Decoder[Prompt] = deriving
-  }
-
-  @newtype case class MaxTokens(value: PosInt)
-  object MaxTokens {
-    implicit val maxTokensEq: Eq[MaxTokens] = deriving
-
-    implicit val maxTokensShow: Show[MaxTokens] = deriving
-
-    implicit val maxTokensEncoder: Encoder[MaxTokens] = deriving
-    implicit val maxTokensDecoder: Decoder[MaxTokens] = deriving
-  }
-
-  type Temperature = Float Refined Interval.Closed[0f, 2f]
-  object Temperature extends RefinedTypeOps.Numeric[Temperature, Float] {
-
-    implicit val temperatureEq: Eq[Temperature] = Eq.fromUniversalEquals
-
-    implicit val temperatureShow: Show[Temperature] = Show.catsShowForFloat.contramap(_.value)
-
-    implicit val temperatureEncoder: Encoder[Temperature] = Encoder.encodeFloat.contramap(_.value)
-    implicit val temperatureDecoder: Decoder[Temperature] = Decoder.decodeFloat.emap(from)
   }
 
   @newtype case class TopP(value: NonNegFloat)
